Work on patrol enemy animation and Ray walk and attack
Improved on the patrol enemy animation. Now the enemy idles at the endpoints and smoothly transitions to walking. Added hit box on the patrol enemy attack.

The player walk animation is triggered when the player moves, instead of constantly running. Player attack animation also added. Works with the attack button on the controller and also keyboard putton "1".